Internationalization: use English source strings as translation keys
Use English source strings as translation keys for strings that appear directly in the source
code. This is in preparation for switching to gettext PO translation files and extracting the
source strings directly from the source code via our vue-i18n gettext extractor.

This does not cover strings that appear not directly in the source code but are shown dynamically
by their key, e.g. translations for error codes, location categories and provider infos.