Internationalization: fix translations for Providers
The changes done in f5939ac8j in provider-assets.ts don't achieve the desired
effect of the actual translation happening in ProviderBanner, because in
production builds, translation keys are replaced with simple numbers, while the
strings that were returned remained as is, and are therefore not valid
translation keys anymore when passed to i18n-t in ProviderBanner.

Now, the translation already happens in prover-assets.ts and is then passed to
ProviderBanner's i18n-t, where the translation will be an unknown translation
key, which will simply be preserved.